+++ plpygis/__init__.py
@@ -1,2 +1,2 @@
 from .geometry import Geometry, Point, LineString, Polygon, MultiPoint, MultiLineString, MultiPolygon, GeometryCollection
-from ._version import __version__, VERSION
+from ._version import __version__